OnnxStack.Converter/FluxDev/requirements.txt
diffusers==0.30.2
sentencepiece==0.2.0
transformers==4.42.4
torch==2.4.1
peft==0.12.0
tabulate==0.9.0
accelerate==0.34.2
protobuf==3.20.3
onnx==1.16.1
onnxruntime-directml==1.19.0
onnxruntime_extensions==0.12.0
olive-ai[directml]==0.6.2